refactor: consolidate theme types and remove legacy theme modules
