+++ libs/common-hooks/src/ThemeProvider.tsx
@@ -4,18 +4,12 @@ import { type PropsWithChildren } from 'react'
 
 import { ThemeProvider as StyledThemeProvider } from 'styled-components/macro'
 
-import { ThemeContext } from './useTheme'
-
-import type { Theme } from './theme'
+import type { DefaultTheme } from 'styled-components/macro'
 
 interface ThemeProviderProps extends PropsWithChildren {
-  theme: Theme
+  theme: DefaultTheme
 }
 
 export function ThemeProvider({ children, theme }: ThemeProviderProps) {
-  return (
-    <ThemeContext.Provider value={theme}>
-      <StyledThemeProvider theme={theme}>{children}</StyledThemeProvider>
-    </ThemeContext.Provider>
-  )
+  return <StyledThemeProvider theme={theme}>{children}</StyledThemeProvider>
 }
